+++ lib/searchkick/query.rb
@@ -505,7 +505,13 @@ module Searchkick
         set_highlights(payload, fields) if options[:highlight]
 
         # timeout shortly after client times out
-        payload[:timeout] ||= "#{Searchkick.search_timeout + 1}s"
+        payload[:timeout] ||=
+          if Searchkick.search_timeout.is_a?(Integer)
+            "#{Searchkick.search_timeout + 1}s"
+          else
+            # ensure at least 1 ms
+            "#{(Searchkick.search_timeout * 1000).ceil}ms"
+          end
 
         # An empty array will cause only the _id and _type for each hit to be returned
         # https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-source-filtering.html